Rename operation factory
